Only write new private files if they don't already exist

+++ bootstrap.sh
@@ -2,7 +2,7 @@
 
 for f in $(cd $HOME/.files/files; find . -type f); do
     bn=$(basename "$f")
-    if [[ "$bn" == ".example"* ]]; then
+    if [[ "$bn" == ".example"* ]] && ! [[ -e $HOME/${bn#.example} ]]; then
         cp "$HOME/.files/files/$f" $HOME/${bn#.example}
         vim $HOME/${bn#.example}
         chmod 600 $HOME/${bn#.example}
